roles

+++ BpmnJava/src/main/java/com/jts/BpmnJava/token/TokenRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     @Query("SELECT t FROM Token t INNER JOIN User u ON t.users.id = u.id " +
             "WHERE u.id = :id AND (t.expired = false OR t.revoked = false)")
-    List<Token> findAllValidTokenByUser(@Param("id") Integer id);
+    List<Token> findAllValidTokenByUser(@Param("id") Long id);
 
     Optional<Token> findByToken(String token);
 }
